pyproject.toml
[project]
name = "travel-agent-api"
version = "0.1.0"
description = ""
authors = [
    {name = "Gabriele Bonazza",email = "gabriele.bonazza92@gmail.com"}
]
readme = "README.md"
requires-python = "<4.0,>=3.9.0"
dependencies = [
    "fastapi (>=0.115.12,<0.116.0)",
    "uvicorn (>=0.34.0,<0.35.0)",
    "openai (>=1.70.0,<2.0.0)",
    "langchain-openai (>=0.3.11,<0.4.0)",
    "langchain-community (>=0.3.20,<0.4.0)",
    "langgraph (>=0.3.22,<0.4.0)",
    "pydantic (>=2.11.1,<3.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "google-search-results (>=2.4.2,<3.0.0)"
]

[tool.poetry]
packages = [{include = "travel_agent_api", from = "src"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
